_1-core-repository/requirements.txt
absl-py==0.15.0
astroid==2.5
astunparse==1.6.3
attrs==21.2.0
cachetools==5.0.0
certifi==2021.10.8
charset-normalizer==2.0.10
click==8.0.3
configparser==5.1.0
cycler==0.10.0
docker-pycreds==0.4.0
flatbuffers==1.12
fonttools==4.29.1
gast==0.4.0
gitdb==4.0.9
GitPython==3.1.24
google-auth==2.5.0
google-auth-oauthlib==0.4.6
google-pasta==0.2.0
grpcio==1.34.1
h5py==3.1.0
idna==3.3
importlib-metadata==4.10.1
isort==5.10.1
joblib==1.1.0
keras-nightly==2.5.0.dev2021032900
Keras-Preprocessing==1.1.2
lazy-object-proxy==1.7.1
Markdown==3.3.6
matplotlib==3.5.1
mccabe==0.6.1
mediapipe==0.8.8
numpy==1.19.5
oauthlib==3.1.1
opencv-contrib-python==4.5.3.56
opt-einsum==3.3.0
packaging==21.3
pandas==1.3.4
pathtools==0.1.2
Pillow==8.3.2
promise==2.3
protobuf==3.19.3
psutil==5.8.0
pyasn1==0.4.8
pyasn1-modules==0.2.8
pylint==2.6.0
pyparsing==2.4.7
python-dateutil==2.8.2
pytz==2021.3
PyYAML==6.0
requests==2.27.1
requests-oauthlib==1.3.0
rsa==4.8
scikit-learn==1.0.1
scipy==1.7.1
seaborn==0.11.2
sentry-sdk==1.4.3
shortuuid==1.0.1
six==1.15.0
smmap==5.0.0
subprocess32==3.5.4
tensorboard==2.8.0
tensorboard-data-server==0.6.1
tensorboard-plugin-wit==1.8.1
tensorflow-estimator==2.5.0
tensorflow-macos==2.5.0
termcolor==1.1.0
threadpoolctl==3.0.0
toml==0.10.2
tqdm==4.62.3
typing-extensions==3.7.4.3
urllib3==1.26.8
wandb==0.12.6
Werkzeug==2.0.2
wrapt==1.12.1
yaspin==2.1.0
zipp==3.7.0
